requirements.txt
boto3
flask
Flask-Session
python-dotenv
